api/requirements.txt
annotated-types
anyio
argon2
argon2-cffi
argon2-cffi-bindings
async-timeout
asyncpg
certifi
cffi
charset-normalizer
click
contourpy
cycler
exceptiongroup
fastapi
filelock
fonttools
fsspec
h11
idna
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python-headless
packaging
pandas
passlib
pillow
psutil
py-cpuinfo
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
pymssql
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
python-multipart
requests
scipy
seaborn
six
sniffio
starlette
sympy
torch
torchvision
tqdm
typing_extensions
tzdata
ultralytics
ultralytics-thop
urllib3
uvicorn
